Using securityfs filesystem
Instead of using a custom trusteesfs that was typically
mounted at /etc/trustees or otherwise, we are using securityfs
which is already part of the kernel.  This is supposed
to be mounted at /sys/kernel/security (the mountpoint already
will exist), but settrustees will find it regardless of where
it is mounted.

Removed the option -t from settrustees for specifying mount point.
If it isn't picked up automatically, this is a bug and should be
reported.

Also adding NEWS file as primary communication for important
changes.