Redid trustee hashing code
The old hashing code was a carryover from the original trustees.

Formerly there was a giant array of trustee_hash_element's that acted
as the hash table.  When a new trustee was added, a hole had to be
found in the hash table to put the new trustee.  When the hash table
got fairly full, a new larger table was created and everything was
readded to the hash table.  This required complicated code and
delicate locking code to keep everything from blowing up while
it was rehashing.

The new system uses a preset number (default 256) of linked lists.
When there is a collision, the element gets put in the linked list.
This is a similar method used in several other place in the kernel
including PID hashes.

In order to support those with many, many more trustee paths, a module
parameter has been added (hash_size).  Something like:

modprobe trustees hash_size=1024

should boot up trustees with a hash size of 1024.  The hash isn't
too heavy, basically each element of the hash adds one pointer.  So
on a 32 bit machine, 256 element hash table only adds about 1kilobyte
of data.  In exchange the actual trustee structure increased in size
by 2 * sizeof(void *) - 1 * sizeof(unsigned int).

Unless there are severe performance hits, the simplicity of the code
should outweigh any minor performance changes.