Rework the rename logic just a bit
It is simple, we should only be looking at the trustees that
affect the file at its old location and at its new location.  If
we are renaming a directory we should also make sure no trustees
are set inside of this directory.  The latter point is not handled
yet, but this at least cleans up the handling of the former.