Change order of initialization/deinitialization
Prior to this patch, trustees initialized the security hooks,
and then the filesystem hooks, and then the global variables.
Occasionally a permissions check would sneak in between the
initialization of the security subsystem and the initialization
of global variables causing an OOPS.  Changing this order to
globals, then filesystem, then security hooks should fix any sort
of issues.