Making kernel component of trustees a real fork of linux
Trustees is now maintained as an actual fork of the linux
kernel.  This is necessary due to the following commit on
linus's git tree which showed up on the 2.6.24 -rc kernels:

------------------------------------------------------------
commit 20510f2f4e2dabb0ff6c13901807627ec9452f98
Author: James Morris <jmorris@namei.org>
Date:   Tue Oct 16 23:31:32 2007 -0700

    security: Convert LSM into a static interface

	Convert LSM into a static interface, as the ability to
	unload a security module is not required by in-tree users
	and potentially complicates the overall security
	architecture.

	Needlessly exported LSM symbols have been unexported, to
	help reduce API abuse.

	Parameters for the capability and root_plug modules are now
	specified at boot.

	The SECURITY_FRAMEWORK_VERSION macro has also been removed.

	In a nutshell, there is no safe way to unload an LSM.  The
	modular interface is thus unecessary and broken
	infrastructure.  It is used only by out-of-tree modules,
	which are often binary-only, illegal, abusive of the API and
	dangerous, e.g.  silently re-vectoring SELinux.
------------------------------------------------------------

As a result, I've placed the new kernel repository to

	git://git.aeruder.net/trustees-kernel.git
	http://git.aeruder.net/trustees-kernel.git

There are several branches in the trustees-kernel.git repository.

mainline
2.6.20-maint
2.6.21-maint
etc...

mainline is periodically updated with the latest vanilla kernel
sources (2.6.xx, 2.6.xx-rcz). 2.6.23-maint would include the
2.6.23.yy kernel releases.  In almost all cases you'd want to
use one of the 2.6.xx-maint branches.

Pain in the neck? Yes.... but at least git makes merging the
changes from the main kernel incredibly easy.